% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{A}
\alias{A}
\title{Return attribute of object}
\usage{
A(obj, attrName)
}
\arguments{
\item{obj}{Object or list of objects}

\item{attrName}{Attribute name as character}
}
\value{
Attribute of object
}
\description{
Return attribute of object
}
