% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{Bond}
\alias{Bond}
\title{Constructor of simple fixed income bond object}
\usage{
Bond(file = NA, dateFormat = "mdy", sep = ",")
}
\arguments{
\item{file}{Name of csv file that contains bond attributes. Heads should have names of attributes}

\item{dateFormat}{File date format from lubridate package i.e. "dmy", "mdy" etc}

\item{sep}{Separator i.e. comma, semicolon or else}
}
\value{
Bond object
}
\description{
Attributes:
$name - name as you wish
$isin - isin code
$currency - currency i.e. USD, RUB...
$faceAmount - face amount
$couponFreq - coupon frequency i.e. 2 for semiannual etc
$issueDate - issue date
$formula - yield formular where STD stands for standard convention, OFZ stands for OFZ convention
$dayCounter - day counter from dayCounterList
$cfactor - conversion factor for T-bonds-notes for T-futures
$couponDates - vector of coupon dates
$couponAmounts - vector of coupon amounts
}
