% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{buildSchedule.FIBond}
\alias{buildSchedule.FIBond}
\title{Return FIBond object with specified standard payment schedule}
\usage{
buildSchedule.FIBond(bond, rate, issueDate, nper, freq = "half",
  face = 1000)
}
\arguments{
\item{bond}{FIBond object}

\item{rate}{Coupon rate}

\item{issueDate}{Issue date}

\item{nper}{Number of coupon periods}

\item{freq}{Frequency of payments: "month", "quarter", "half", "year"}

\item{face}{Face value}
}
\value{
FIBond object with specified schedule
}
\description{
Return FIBond object with specified standard payment schedule
}
