% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{checkTicker.TFutures}
\alias{checkTicker.TFutures}
\title{Check TFutures ticker}
\usage{
checkTicker.TFutures(ticker)
}
\arguments{
\item{ticker}{Ticker}
}
\value{
Ticker if it is a ticker, NA otherwise
}
\description{
Check TFutures ticker
}
