% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{countDays}
\alias{countDays}
\title{Return number of days between two dates for a given day counter}
\usage{
countDays(startDate, endDate, dayCounter)
}
\arguments{
\item{startDate}{Start date}

\item{endDate}{End date}

\item{DAY_COUNTER}{Day counter}
}
\value{
Number of days between two dates for a given day counter
}
\description{
Return number of days between two dates for a given day counter
}
