% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{counterName}
\alias{counterName}
\title{Return name of day counter}
\usage{
counterName(counter)
}
\arguments{
\item{counter}{Counter as a number (can be a vector)}
}
\value{
Name
}
\description{
Return name of day counter
}
\examples{
counter(0)
}
