% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\docType{data}
\name{DAY_COUNTER}
\alias{DAY_COUNTER}
\title{List of RQuantLib day counters}
\format{An object of class \code{list} of length 14.}
\usage{
DAY_COUNTER
}
\description{
List of RQuantLib day counters
}
\keyword{datasets}
