% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\docType{data}
\name{dayCounter}
\alias{dayCounter}
\title{List of RQuantLib day counters}
\format{An object of class \code{list} of length 14.}
\usage{
dayCounter
}
\description{
List of RQuantLib day counters
}
\keyword{datasets}
