% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{demoTFutModelError}
\alias{demoTFutModelError}
\title{Chart historical TFutures pricing model errors for XXU7 contracts}
\usage{
demoTFutModelError(ticker, ois3 = FALSE, ois4 = FALSE)
}
\arguments{
\item{ticker}{Ticker or its number from futDataList like TYU7 etc}

\item{ois3}{Chart errors for USD OIS 3M as a term repo rate if TRUE}

\item{ois4}{Chart errors for USD OIS 4M as a term repo rate if TRUE}
}
\value{
Chart of TFutures pricing model errors
}
\description{
Chart historical TFutures pricing model errors for XXU7 contracts
}
