% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{demoTFutSensToRP}
\alias{demoTFutSensToRP}
\title{Chart ageing of PVBP relative to term repo rate for XXU7 contracts}
\usage{
demoTFutSensToRP(ticker)
}
\arguments{
\item{ticker}{Ticker like TYU7 etc}
}
\value{
Chart of PVBP relative to term repo rate over time
}
\description{
Chart ageing of PVBP relative to term repo rate for XXU7 contracts
}
