% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{dir.FIBond}
\alias{dir.FIBond}
\title{Show attributes of all bonds from a given folder}
\usage{
dir.FIBond(folder = FOLDER_FIBONDS_US_GOV, attr = c("name", "isin", "risk",
  "issueDate", "maturity"))
}
\arguments{
\item{folder}{Folder}

\item{attr}{Vector of attributes' names}
}
\value{
Data frame with bonds' attributes
}
\description{
Show attributes of all bonds from a given folder
}
