% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{dir.TFutures}
\alias{dir.TFutures}
\title{Show attributes of all ctd bonds in FOLDER_TFUTURES}
\usage{
dir.TFutures(attr = c("name", "isin", "issueDate", "maturity"))
}
\arguments{
\item{attr}{Vector of attributes' names}
}
\value{
Data frame with bonds' attributes
}
\description{
Show attributes of all ctd bonds in FOLDER_TFUTURES
}
