% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{disassembleDate}
\alias{disassembleDate}
\title{Disassemble date object to three numericals - day, month, year}
\usage{
disassembleDate(date)
}
\arguments{
\item{date}{Date object}
}
\value{
list of numericals $day, $month, $year
}
\description{
Disassemble date object to three numericals - day, month, year
}
