% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{firstBizDay}
\alias{firstBizDay}
\title{Return date of first business day for a given month and year}
\usage{
firstBizDay(month, year, calendar = "WeekendsOnly")
}
\arguments{
\item{month}{Month number}

\item{year}{Year}

\item{calendar}{RQuantLib calendar name}
}
\value{
Date of first business day of a given month and year
}
\description{
Return date of first business day for a given month and year
}
