% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{firstDay}
\alias{firstDay}
\title{Return date of first day for a given month and year}
\usage{
firstDay(month, year)
}
\arguments{
\item{month}{Month number}

\item{year}{Year}
}
\value{
Date of first day for a given month and year
}
\description{
Return date of first day for a given month and year
}
