% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getAccrued.FIBond}
\alias{getAccrued.FIBond}
\title{Calculate accrued interest for FIBond object}
\usage{
getAccrued.FIBond(bond, settleDate = nextBizDay())
}
\arguments{
\item{bond}{FIBond object}

\item{settleDate}{Calculation date (can be a vector)}
}
\value{
Accrued interest
}
\description{
Calculate accrued interest for FIBond object
}
