% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getAccrued}
\alias{getAccrued}
\title{Calculate accrued interest for Bond object}
\usage{
getAccrued(bond, settleDate = nextBizDay())
}
\arguments{
\item{bond}{Bond object}

\item{settleDate}{Calculation date}
}
\value{
Accrued interest
}
\description{
Calculate accrued interest for Bond object
}
