% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getAccruedForPrice.FIBond}
\alias{getAccruedForPrice.FIBond}
\title{Calculate accrued interest in percentage of current face where 1 is 100\%}
\usage{
getAccruedForPrice.FIBond(bond, settleDate = nextBizDay())
}
\arguments{
\item{bond}{FIBond object}

\item{settleDate}{Calculation date (can be a vector)}
}
\value{
Accrued interest in percentage of current face where 1 is 100\%
}
\description{
Calculate accrued interest in percentage of current face where 1 is 100\%
}
