% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{getAccruedPer100}
\alias{getAccruedPer100}
\title{Calculate accrued interest for Bond object per 100 of face}
\usage{
getAccruedPer100(bond, settleDate = nextBizDay())
}
\arguments{
\item{bond}{Bond object}

\item{settleDate}{Calculation date}
}
\value{
Accrued interest per 100 of face
}
\description{
Calculate accrued interest for Bond object per 100 of face
}
