% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{getAccruedPrice.FIBond}
\alias{getAccruedPrice.FIBond}
\title{Calculate accrued interest in percentage of current face}
\usage{
getAccruedPrice.FIBond(bond, settleDate = nextBizDay())
}
\arguments{
\item{bond}{FIBond object}

\item{settleDate}{Calculation date (can be a vector)}
}
\value{
Accrued interest in percentage of current face
}
\description{
Calculate accrued interest in percentage of current face
}
