% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{getCarry.FIBond}
\alias{getCarry.FIBond}
\title{Calculate carry for FIBond object in percentage of current face}
\usage{
getCarry.FIBond(bond, price, settleDate1, settleDate2, repoRate)
}
\arguments{
\item{bond}{FIBond object (can be a list)}

\item{price}{FIBond clean price in percentage}

\item{settleDate1}{Start date}

\item{settleDate2}{End date}

\item{repoRate}{Funding term repo rate in percentage}
}
\value{
Carry for FIBond object in percentage of current face
}
\description{
Calculate carry for FIBond object in percentage of current face
}
