% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getCarry.TFututes}
\alias{getCarry.TFututes}
\title{Return percentage carry of FIBond till delivery date of TFutures}
\usage{
getCarry.TFututes(fut, bondPrice, repoRate, tradeDate = Sys.Date(),
  bond = takeCTD.TFutures(fut))
}
\arguments{
\item{fut}{TFutures object (can be a list)}

\item{bondPrice}{FIBond price in percentage (can be a vector)}

\item{repoRate}{FIBond repo rate in percentage (can be a vector)}

\item{tradeDate}{Trade date (can be a vector)}

\item{bond}{FIBond object}
}
\value{
Carry in percentage of FIBond face
}
\description{
Return percentage carry of FIBond till delivery date of TFutures
}
