% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getCarryForPrice.FIBond}
\alias{getCarryForPrice.FIBond}
\title{Calculate carry for FIBond object in percentage of current face where 1 is 100\%}
\usage{
getCarryForPrice.FIBond(bond, price, settleDate1, settleDate2, repoRate)
}
\arguments{
\item{bond}{FIBond object}

\item{price}{FIBond clean price where 1 is 100\% (can be a vector)}

\item{settleDate1}{Start date (can be a vector)}

\item{settleDate2}{End date (can be a vector)}

\item{repoRate}{Funding term repo rate (can be a vector)}
}
\value{
Carry for FIBond object in percentage of current face where 1 is 100\%
}
\description{
Calculate carry for FIBond object in percentage of current face where 1 is 100\%
}
