% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getCarryPer100}
\alias{getCarryPer100}
\title{Calculate carry for Bond object for 100 face}
\usage{
getCarryPer100(bond, price, settleDate1, settleDate2, repoRate)
}
\arguments{
\item{bond}{Bond object}

\item{price}{Clean price of Bond object for 100 face}

\item{settleDate1}{Start date}

\item{settleDate2}{End date}

\item{repoRate}{Funding term repo rate}
}
\value{
Carry for Bond object
}
\description{
Calculate carry for Bond object for 100 face
}
