% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getCFactor.TFutures}
\alias{getCFactor.TFutures}
\title{Return covnersion factor for given FIBond and TFutures}
\usage{
getCFactor.TFutures(bond, ticker, decade = "auto")
}
\arguments{
\item{bond}{FIBond object}

\item{ticker}{Ticker of TFutures}

\item{decade}{Decade that can be "auto" - default value, "pres" - present decade,
"prev" - previous decade, "next" - next decade}
}
\value{
Conversion factor
}
\description{
Return covnersion factor for given FIBond and TFutures
}
