% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futures.R
\name{getCode.Futures}
\alias{getCode.Futures}
\title{Take contract type code from futures ticker}
\usage{
getCode.Futures(ticker)
}
\arguments{
\item{ticker}{Ticker}
}
\value{
Contract type code
}
\description{
Take contract type code from futures ticker
}
