% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{getCoupon.FIBond}
\alias{getCoupon.FIBond}
\title{Return current coupon amount}
\usage{
getCoupon.FIBond(bond, settleDate = nextBizDay())
}
\arguments{
\item{bond}{FIBond object}

\item{settleDate}{Calculation date}
}
\value{
Current coupon amount
}
\description{
Return current coupon amount
}
