% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getCouponTime}
\alias{getCouponTime}
\title{Calculate coupon time i.e. days passed over days in coupon period for Bond object}
\usage{
getCouponTime(bond, settleDate = nextBizDay())
}
\arguments{
\item{bond}{Bond object}

\item{settleDate}{Calculation date (can be a vector)}
}
\value{
Coupon time for Bond object
}
\description{
Calculate coupon time i.e. days passed over days in coupon period for Bond object
}
