% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getCurrentFace.FIBond}
\alias{getCurrentFace.FIBond}
\title{Return current face amount}
\usage{
getCurrentFace.FIBond(bond, settleDate = nextBizDay())
}
\arguments{
\item{bond}{FIBond object}

\item{settleDate}{Calculation date (can be a vector)}
}
\value{
Current face amount
}
\description{
Return current face amount
}
