% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{getDeliveryDate.TFutures}
\alias{getDeliveryDate.TFutures}
\title{Return model delivery date by TFutures ticker}
\usage{
getDeliveryDate.TFutures(ticker, decade = "auto")
}
\arguments{
\item{ticker}{Ticker}

\item{decade}{decade that can be "auto" - default value, "pres" - present decade,
"prev" - previous decade, "next" - next decade}
}
\value{
Delivery date
}
\description{
Return model delivery date by TFutures ticker
}
