% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{getDur.FIBond}
\alias{getDur.FIBond}
\title{Return modified duration of given FIBond}
\usage{
getDur.FIBond(bond, price, settleDate = nextBizDay())
}
\arguments{
\item{bond}{FIBond object}

\item{price}{FIBond clean price in percentage}

\item{settleDate}{Calculation date}
}
\value{
Modified duration of given FIBond
}
\description{
Return modified duration of given FIBond
}
