% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getEligibles.TFutures}
\alias{getEligibles.TFutures}
\title{Return ISINs of eligible FIBonds for giben ticker}
\usage{
getEligibles.TFutures(ticker)
}
\arguments{
\item{ticker}{Ticker of TFutures}

\item{folder}{FIBond's folder}
}
\value{
Vector of ISINs
}
\description{
Return ISINs of eligible FIBonds for giben ticker
}
