% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getFirstDay.TFutures}
\alias{getFirstDay.TFutures}
\title{Return first bisness day of delivery month by TFutures ticker}
\usage{
getFirstDay.TFutures(ticker, decade = "auto")
}
\arguments{
\item{ticker}{Ticker}

\item{decade}{Decade that can be "auto" - default value, "pres" - present decade,
"prev" - previous decade, "next" - next decade}
}
\value{
First business day
}
\description{
Return first bisness day of delivery month by TFutures ticker
}
