% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getFirstIntention.TFutures}
\alias{getFirstIntention.TFutures}
\title{Return first intention date by TFutures ticker}
\usage{
getFirstIntention.TFutures(ticker, decade = "auto")
}
\arguments{
\item{ticker}{Ticker}

\item{decade}{Decade that can be "auto" - default value, "pres" - present decade,
"prev" - previous decade, "next" - next decade}
}
\value{
First intention date
}
\description{
Return first intention date by TFutures ticker
}
