% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{getImpliedRepoRate}
\alias{getImpliedRepoRate}
\title{Calculate implied repo rate for TFutures object}
\usage{
getImpliedRepoRate(fut, futPrice, ctdPrice, tradeDate = Sys.Date())
}
\arguments{
\item{fut}{TFutures object}

\item{futPrice}{TFutures price for 100 notional}

\item{ctdPrice}{CTD bond clean price}

\item{tradeDate}{Calculation date}
}
\value{
Implied repo rate for TFutures object
}
\description{
Calculate implied repo rate for TFutures object
}
