% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getIRP.TFututes}
\alias{getIRP.TFututes}
\title{Calculate implied repo rate for TFutures object}
\usage{
getIRP.TFututes(fut, futPrice, ctdPrice, tradeDate = Sys.Date())
}
\arguments{
\item{fut}{TFutures object}

\item{futPrice}{TFutures price (can be a vector)}

\item{ctdPrice}{CTD bond clean price (can be a vector)}

\item{tradeDate}{Calculation date (can be a vector)}
}
\value{
Implied repo rate for TFutures object
}
\description{
Calculate implied repo rate for TFutures object
}
