% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getLastTradingDate.TFutures}
\alias{getLastTradingDate.TFutures}
\title{Return last trading date by TFutures ticker}
\usage{
getLastTradingDate.TFutures(ticker, decade = "auto")
}
\arguments{
\item{ticker}{Ticker}

\item{decade}{Decade that can be "auto" - default value, "pres" - present decade,
"prev" - previous decade, "next" - next decade}
}
\value{
Last trading date
}
\description{
Return last trading date by TFutures ticker
}
