% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Futures.R
\name{getMonth.Futures}
\alias{getMonth.Futures}
\title{Take expiration month number from futures ticker}
\usage{
getMonth.Futures(ticker)
}
\arguments{
\item{ticker}{Ticker}
}
\value{
Expiration month number
}
\description{
Take expiration month number from futures ticker
}
