% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{getName.TFutures}
\alias{getName.TFutures}
\title{Return contract name by TFutures ticker}
\usage{
getName.TFutures(ticker, decade = "auto")
}
\arguments{
\item{ticker}{Ticker}

\item{decade}{decade that can be "auto" - default value, "pres" - present decade,
"prev" - previous decade, "next" - next decade}
}
\value{
Contract name
}
\description{
Return contract name by TFutures ticker
}
