% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getNotional.TFutures}
\alias{getNotional.TFutures}
\title{Return notional amount by TFutures ticker}
\usage{
getNotional.TFutures(ticker)
}
\arguments{
\item{ticker}{Ticker (can be a vector)}
}
\value{
Notional amount
}
\description{
Return notional amount by TFutures ticker
}
