% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{getPrice.FIBond}
\alias{getPrice.FIBond}
\title{Calculate clean price of FIBond}
\usage{
getPrice.FIBond(bond, yield, settleDate = nextBizDay())
}
\arguments{
\item{bond}{FIBond object (can be a list)}

\item{yield}{FIBond yield in percentage (can be a vector)}

\item{settleDate}{Calculation date (can be a vector)}
}
\value{
Clean price of FIBond object in percentage
}
\description{
Calculate clean price of FIBond
}
