% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getPrice.TFutures}
\alias{getPrice.TFutures}
\title{Calculate model price of TFutures object}
\usage{
getPrice.TFutures(fut, ctdPrice, repoRate, tradeDate = Sys.Date())
}
\arguments{
\item{fut}{TFutures object}

\item{ctdPrice}{CTD bond clean price in percentage (can be a vector)}

\item{repoRate}{Term CTD repo rate in percentage (can be a vector)}

\item{tradeDate}{Calculation date (can be a vector)}
}
\value{
Model price of TFutures object in percentage of notional
}
\description{
Calculate model price of TFutures object
}
