% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getPriceOfBond}
\alias{getPriceOfBond}
\title{Calculate clean price of Bond object for 100 face}
\usage{
getPriceOfBond(bond, yield, settleDate = nextBizDay())
}
\arguments{
\item{bond}{Bond object}

\item{yield}{Bond yield}

\item{settleDate}{Calculation date}
}
\value{
Clean price of a Bond object
}
\description{
Calculate clean price of Bond object for 100 face
}
