% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{getPriceOfTFutures}
\alias{getPriceOfTFutures}
\title{Calculate model price of TFutures object for 100 notional}
\usage{
getPriceOfTFutures(fut, ctdPrice, repoRate, tradeDate = Sys.Date())
}
\arguments{
\item{fut}{TFutures object}

\item{ctdPrice}{CTD bond clean price}

\item{repoRate}{Term CTD repo rate}

\item{tradeDate}{Calculation date}
}
\value{
Model price of TFutures object
}
\description{
Calculate model price of TFutures object for 100 notional
}
