% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVM.R
\name{getPV.TVM}
\alias{getPV.TVM}
\title{Return present value of periodic payments}
\usage{
getPV.TVM(rate, nper, pmt, fv, lessAccrued = FALSE)
}
\arguments{
\item{rate}{Interest rate in percentage}

\item{nper}{Number of periods (can be not a whole number)}

\item{pmt}{Payment (all payments are equal)}

\item{fv}{Future value}

\item{lessAccrued}{TRUE for deducting accrued, FALSE otherwise}
}
\value{
Present value of periodic payments
}
\description{
Return present value of periodic payments
}
