% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getPVBPofBond}
\alias{getPVBPofBond}
\title{Calculate PVBP of Bond object}
\usage{
getPVBPofBond(bond, price, settleDate = nextBizDay())
}
\arguments{
\item{bond}{Bond object}

\item{price}{Clean price of Bond object for 100 face}

\item{settleDate}{Calculation date}
}
\value{
PVBP of Bond object
}
\description{
Calculate PVBP of Bond object
}
