% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getPVBPRP.TFutures}
\alias{getPVBPRP.TFutures}
\title{Calculate PVBP of TFutures object relative to CTD repo rate change}
\usage{
getPVBPRP.TFutures(fut, ctdPrice, repoRate, tradeDate = Sys.Date())
}
\arguments{
\item{fut}{TFutures object}

\item{ctdPrice}{CTD bond clean price in percentage}

\item{repoRate}{CTD repo rate in percentage}

\item{tradeDate}{Calculation date}
}
\value{
PVBP of TFutures object relative to CTD repo rate change in percentage
}
\description{
Calculate PVBP of TFutures object relative to CTD repo rate change
}
