% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{getPVBPRPofTFutures}
\alias{getPVBPRPofTFutures}
\title{Calculate PVBP of TFutures object relative to CTD repo rate change}
\usage{
getPVBPRPofTFutures(fut, ctdPrice, repoRate, tradeDate = Sys.Date())
}
\arguments{
\item{fut}{TFutures object}

\item{ctdPrice}{CTD bond clean price for 100 face}

\item{repoRate}{CTD repo rate}

\item{tradeDate}{Calculation date}
}
\value{
PVBP of TFutures object relative to CTD repo rate change
}
\description{
Calculate PVBP of TFutures object relative to CTD repo rate change
}
