% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{getRate.FIBond}
\alias{getRate.FIBond}
\title{Return coupon rate for current period}
\usage{
getRate.FIBond(bond, settleDate = nextBizDay())
}
\arguments{
\item{bond}{FIBond object}

\item{settleDate}{Calculation date}
}
\value{
Coupon rate for current period in percentage
}
\description{
Return coupon rate for current period
}
