% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{getSchedule}
\alias{getSchedule}
\title{Return vector of dates with given period from given start date}
\usage{
getSchedule(startDate, nper, period = "month")
}
\arguments{
\item{startDate}{Start date}

\item{nper}{Number of periods}

\item{period}{Period - "month", "quarter", "half", "year"}
}
\value{
Vector of dates with given period from given start date
}
\description{
Return vector of dates with given period from given start date
}
