% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getTFutModelAnalytics}
\alias{getTFutModelAnalytics}
\title{Return historical data analytics for TFuture pricing model}
\usage{
getTFutModelAnalytics(fut, histFile, futField = "futPrices",
  ctdField = "ctdPrices", repoField = "repoRates", dateFormat = "ymd",
  sep = ",")
}
\arguments{
\item{fut}{TFutures object}

\item{histFile}{csv file that contains historical futures prices, CTD prices, repo rates}

\item{futField}{Head of futures prices column in csv file}

\item{ctdField}{Head of CTD prices column in csv file}

\item{repoField}{Head of CTD repo rates column in csv file}

\item{dateFormat}{File date format from lubridate package i.e. "dmy", "mdy" etc}

\item{sep}{Separator i.e. comma, semicolon or else}
}
\value{
Data frame with impliedRates, modelPrices, errors32nds
}
\description{
Return historical data analytics for TFuture pricing model
}
