% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{getTime.FIBond}
\alias{getTime.FIBond}
\title{Calculate coupon time i.e. days passed over days in coupon period for FIBond object}
\usage{
getTime.FIBond(bond, settleDate = nextBizDay())
}
\arguments{
\item{bond}{FIBond object}

\item{settleDate}{Calculation date}
}
\value{
List of $period (days), $passed (days), $time (ratio) values
}
\description{
Calculate coupon time i.e. days passed over days in coupon period for FIBond object
}
