% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getType.TFutures}
\alias{getType.TFutures}
\title{Return contract type by TFutures ticker}
\usage{
getType.TFutures(ticker)
}
\arguments{
\item{ticker}{Ticker}
}
\value{
Contract type
}
\description{
Return contract type by TFutures ticker
}
