% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurOption.R
\name{getValue.EurOption}
\alias{getValue.EurOption}
\title{Get value of european option
Black-Scholes}
\usage{
getValue.EurOption(option, price)
}
\arguments{
\item{option}{option object}

\item{price}{price of underlying (can be a vector)}
}
\value{
value of option
}
\description{
Get value of european option
Black-Scholes
}
