% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getValue.TFutures}
\alias{getValue.TFutures}
\title{Return value of TFutures position}
\usage{
getValue.TFutures(fut, futPrice, settlePrice, numOfContracts = 1)
}
\arguments{
\item{fut}{TFutures object}

\item{futPrice}{TFutures price in percentage}

\item{settlePrice}{TFutures settlement price in percentage}

\item{numOfContracts}{Numbe of contracts (negative is short)}
}
\value{
Value in dollars
}
\description{
Return value of TFutures position
}
