% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getValueOfBond}
\alias{getValueOfBond}
\title{Calculate value of Bond object}
\usage{
getValueOfBond(bond, yield, settleDate = nextBizDay())
}
\arguments{
\item{bond}{Bond object}

\item{yield}{Bond yield}

\item{settleDate}{Calculation date}
}
\value{
Dirty value of bond object
}
\description{
Calculate value of Bond object
}
