% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurOption.R
\name{getValueOfEurOption}
\alias{getValueOfEurOption}
\title{Get value of european option
Black-Scholes}
\usage{
getValueOfEurOption(option, price)
}
\arguments{
\item{option}{option object}

\item{price}{price of underlying}
}
\value{
value of option
}
\description{
Get value of european option
Black-Scholes
}
