% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Futures.R
\name{getYear.Futures}
\alias{getYear.Futures}
\title{Take contract's year from futures ticker}
\usage{
getYear.Futures(ticker, decade = "auto")
}
\arguments{
\item{ticker}{Ticker}

\item{decade}{Decade that can be "auto" - default value, "pres" - present decade,
"prev" - previous decade, "next" - next decade}
}
\value{
Contract's year
}
\description{
Take contract's year from futures ticker
}
