% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{getYield.FIBond}
\alias{getYield.FIBond}
\title{Calculate yield of FIBond object}
\usage{
getYield.FIBond(bond, price, settleDate = nextBizDay(), digits = 4,
  yieldRange = c(-90, 90))
}
\arguments{
\item{bond}{FIBond object (can be a list)}

\item{price}{FIBond clean price in percentage (can be a vector)}

\item{settleDate}{Calculation date (can be a vector)}

\item{digits}{Accuracy as a number of digits after point}
}
\value{
FIBond yield in percentage
}
\description{
Calculate yield of FIBond object
}
