% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getYieldOfBond}
\alias{getYieldOfBond}
\title{Calculate yield of Bond object}
\usage{
getYieldOfBond(bond, price, settleDate = nextBizDay(), digits = 4)
}
\arguments{
\item{bond}{Bond object}

\item{price}{Bond clean price for 100 face}

\item{settleDate}{Calculation date}

\item{digits}{Accuracy as a number of digits after point}
}
\value{
Bond yield
}
\description{
Calculate yield of Bond object
}
