% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{isBizDay}
\alias{isBizDay}
\title{Is given date a business day}
\usage{
isBizDay(calendar, date)
}
\arguments{
\item{calendar}{Calendar name from RQuantLib}

\item{date}{Date}
}
\value{
TRUE if business day, FALSE otherwise
}
\description{
Is given date a business day
}
