% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{isEligible.TFutures}
\alias{isEligible.TFutures}
\title{Return TRUE if FIBond is eligible for given ticker}
\usage{
isEligible.TFutures(issueDate, maturity, ticker, decade = "auto")
}
\arguments{
\item{issueDate}{Issue date of FIBond}

\item{maturity}{Maturity date of FIBond}

\item{ticker}{Ticker of TFutures}

\item{decade}{Decade that can be "auto" - default value, "pres" - present decade,
"prev" - previous decade, "next" - next decade}
}
\value{
TRUE if FIBond is form basket of TFutures, otherwise FALSE
}
\description{
Return TRUE if FIBond is eligible for given ticker
}
