% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{isEligible2.TFutures}
\alias{isEligible2.TFutures}
\title{Return TRUE if FIBond is eligible for given ticker}
\usage{
isEligible2.TFutures(bond, ticker, decade = "auto")
}
\arguments{
\item{bond}{FIBond object}

\item{ticker}{Ticker of TFutures}

\item{decade}{Decade that can be "auto" - default value, "pres" - present decade,
"prev" - previous decade, "next" - next decade}
}
\value{
TRUE if FIBond is form basket of TFutures, otherwise FALSE
}
\description{
Return TRUE if FIBond is eligible for given ticker
}
