% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futures.R
\name{isTicker.Futures}
\alias{isTicker.Futures}
\title{Is Futures ticker}
\usage{
isTicker.Futures(ticker)
}
\arguments{
\item{ticker}{Ticker}
}
\value{
TRUE if it is a ticker, FALSE otherwise
}
\description{
Is Futures ticker
}
