% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{isTicker.TFutures}
\alias{isTicker.TFutures}
\title{Is TFutures ticker}
\usage{
isTicker.TFutures(ticker)
}
\arguments{
\item{ticker}{Ticker}
}
\value{
TRUE if it is a ticker, FALSE otherwise
}
\description{
Is TFutures ticker
}
