% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{L}
\alias{L}
\title{Check that all params have length = 1 or same length > 1}
\usage{
L(...)
}
\arguments{
\item{...}{}
}
\value{
Maximum length of params
}
\description{
Check that all params have length = 1 or same length > 1
}
