% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{lastBizDay}
\alias{lastBizDay}
\title{Return date of last business day for a given month and year}
\usage{
lastBizDay(month, year, calendar = "WeekendsOnly")
}
\arguments{
\item{month}{Month number}

\item{year}{Year}

\item{calendar}{RQuantLib calendar name}
}
\value{
Date of last business day of a given month and year
}
\description{
Return date of last business day for a given month and year
}
