% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futures.R
\name{letter.Futures}
\alias{letter.Futures}
\title{Return month code for a given month number}
\usage{
letter.Futures(month)
}
\arguments{
\item{month}{Month number}
}
\value{
Month code
}
\description{
Return month code for a given month number
}
