% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{loadCTD.TFutures}
\alias{loadCTD.TFutures}
\title{Return FIBond object of CTD for a given TFutures}
\usage{
loadCTD.TFutures(ticker, file = "")
}
\arguments{
\item{ticker}{Ticker of TFutures}

\item{file}{Filename of CTD without extension}
}
\value{
FIBond object of CTD
}
\description{
Return FIBond object of CTD for a given TFutures
}
