% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\name{loadMarket}
\alias{loadMarket}
\title{Load historical market prices and rates}
\usage{
loadMarket(file = "market")
}
\arguments{
\item{file}{File name of csv data file without extension}
}
\value{
Data frame of market data
}
\description{
Load historical market prices and rates
}
