% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Futures.R
\name{month.Futures}
\alias{month.Futures}
\title{Return month number for a given month code}
\usage{
month.Futures(letter)
}
\arguments{
\item{code}{Month code}
}
\value{
Month number
}
\description{
Return month number for a given month code
}
