% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{nextBizDay}
\alias{nextBizDay}
\title{Give next business day using RQuantLib calendars}
\usage{
nextBizDay(date = Sys.Date(), calendar = "WeekendsOnly")
}
\arguments{
\item{date}{Date}

\item{calendar}{RQuantLib calendar name}
}
\value{
Next business day
}
\description{
Give next business day using RQuantLib calendars
}
