% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{nextMonth}
\alias{nextMonth}
\title{Return next month's number and year}
\usage{
nextMonth(month, year, n = 1)
}
\arguments{
\item{month}{Present month's number}

\item{year}{Year}

\item{n}{Number of months to go forward}
}
\value{
List where $month is next month and $year is year of next month
}
\description{
Return next month's number and year
}
