% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{parseDate}
\alias{parseDate}
\title{Return date object}
\usage{
parseDate(date, format)
}
\arguments{
\item{date}{Date as a string}

\item{format}{dmy, mdy, ymd}
}
\value{
Date object
}
\description{
Return date object
}
