% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{plusMonth}
\alias{plusMonth}
\title{Return date n month forward from given date}
\usage{
plusMonth(date, n = 1)
}
\arguments{
\item{date}{Date}

\item{n}{Number of months to go forward}
}
\value{
Date n month forward
}
\description{
Return date n month forward from given date
}
