% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{prevBizDay}
\alias{prevBizDay}
\title{Return previous business day using RQuantLib calendars}
\usage{
prevBizDay(date = Sys.Date(), calendar = "WeekendsOnly")
}
\arguments{
\item{date}{Date}

\item{calendar}{RQuantLib calendar name}
}
\value{
Next business day
}
\description{
Return previous business day using RQuantLib calendars
}
