% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{print.TFutures}
\alias{print.TFutures}
\title{Print TFutures object}
\usage{
\method{print}{TFutures}(fut)
}
\arguments{
\item{bond}{TFutures object}
}
\value{
Info on screen
}
\description{
Print TFutures object
}
