% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{roundSpan}
\alias{roundSpan}
\title{Return span between two dates in whole months, quarters, halfyears, years}
\usage{
roundSpan(startDate, endDate, roundBy = "month")
}
\arguments{
\item{startDate}{Start date}

\item{endDate}{End date}

\item{roundBy}{Round by "month", "quarter", "half", "year"}
}
\value{
Span between two dates
}
\description{
Return span between two dates in whole months, quarters, halfyears, years
}
