% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{roundTerm.TFutures}
\alias{roundTerm.TFutures}
\title{Round bond's term to maturity according to TFutures rules}
\usage{
roundTerm.TFutures(bond, ticker, decade = "auto")
}
\arguments{
\item{bond}{FIBond object}

\item{ticker}{TFutures ticker}

\item{decade}{Decade that can be "auto" - default value, "pres" - present decade,
"prev" - previous decade, "next" - next decade}
}
\value{
Rounded bond's term in years
}
\description{
Round bond's term to maturity according to TFutures rules
}
