% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\name{showFolder.FIBond}
\alias{showFolder.FIBond}
\title{Show prime attributes of all bonds from a given folder}
\usage{
showFolder.FIBond(folder = FIBONDS_FOLDER)
}
\arguments{
\item{folder}{Folder}
}
\value{
Data frame with bonds' attributes
}
\description{
Show prime attributes of all bonds from a given folder
}
