% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{takeCFactor.FIBond}
\alias{takeCFactor.FIBond}
\title{Return cfactor of FIBond object}
\usage{
takeCFactor.FIBond(bond)
}
\arguments{
\item{bond}{FIBond object (can be a list)}
}
\value{
CFactor (can be a vector)
}
\description{
Return cfactor of FIBond object
}
