% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{takeCTD.TFutures}
\alias{takeCTD.TFutures}
\title{Return CTD of TFutures object}
\usage{
takeCTD.TFutures(fut)
}
\arguments{
\item{fut}{TFutures object (can be a list)}
}
\value{
FIBond object of CTD (can be a list)
}
\description{
Return CTD of TFutures object
}
