% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{takeISIN.TFutures}
\alias{takeISIN.TFutures}
\title{Return ISIN of TFutures object}
\usage{
takeISIN.TFutures(fut)
}
\arguments{
\item{fut}{TFutures object (can be a list)}
}
\value{
ISIN as character (can be a vector)
}
\description{
Return ISIN of TFutures object
}
