% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{vectorSwitch}
\alias{vectorSwitch}
\title{Common switch that takes vector as an expression argument - helper.R}
\usage{
vectorSwitch(expression, ...)
}
\arguments{
\item{expression}{Expression}

\item{...}{Cases}
}
\value{
Case that matches expression
}
\description{
Common switch that takes vector as an expression argument - helper.R
}
\examples{
vectorSwitch(c("a","c"), a = 1, b = 2, c = 3, 4)
}
