% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{demoTFutRatesHistory}
\alias{demoTFutRatesHistory}
\title{Chart historical interest rates for some period of 2017 related to XXU7 contracts}
\usage{
demoTFutRatesHistory(histFile = "fval_data/hist_tuu7.csv", ois3 = FALSE,
  ois4 = FALSE)
}
\arguments{
\item{histFile}{History file where function takes interest rate data}

\item{ois3}{Chart USD OIS 3M if TRUE}

\item{ois4}{Chart USD OIS 4M if TRUE}
}
\value{
Chart of historical interest rates
}
\description{
Chart historical interest rates for some period of 2017 related to XXU7 contracts
}
