% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{dir.FIBond}
\alias{dir.FIBond}
\title{Show attributes of all bonds from a given folder}
\usage{
dir.FIBond(attr = c("name", "isin", "risk", "issueDate", "maturity"),
  folder = FIBONDS_FOLDER, showFiles = TRUE)
}
\arguments{
\item{attr}{Vector of attributes' names}

\item{folder}{Folder}
}
\value{
Data frame with bonds' attributes
}
\description{
Show attributes of all bonds from a given folder
}
