% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{getAccruedValue.FIBond}
\alias{getAccruedValue.FIBond}
\title{Calculate accrued interest for FIBond object}
\usage{
getAccruedValue.FIBond(bond, settleDate = nextBizDay())
}
\arguments{
\item{bond}{FIBond object (can be a list)}

\item{settleDate}{Calculation date (can be a vector)}
}
\value{
Accrued interest
}
\description{
Calculate accrued interest for FIBond object
}
