% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getBasis.TFutures}
\alias{getBasis.TFutures}
\title{Return Basis for TFutures object and given FIBond}
\usage{
getBasis.TFutures(fut, futPrice, bondPrice)
}
\arguments{
\item{fut}{TFutures object (can be a list)}

\item{futPrice}{TFutures price in percentage (can be a vector)}

\item{bondPrice}{FIBond price in percentage (can be a vector)}
}
\value{
Basis in percentage of FIBond face
}
\description{
Return Basis for TFutures object and given FIBond
}
