% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getBasket.TFutures}
\alias{getBasket.TFutures}
\title{Return ISINs of FIBonds from basket}
\usage{
getBasket.TFutures(ticker, folder = FOLDER_FIBONDS)
}
\arguments{
\item{ticker}{Ticker of TFutures}

\item{folder}{FIBond's folder}
}
\value{
Vector os ISINs
}
\description{
Return ISINs of FIBonds from basket
}
