% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getBondTerm.TFutures}
\alias{getBondTerm.TFutures}
\title{Return rounded bond's term to maturity according to TFutures rules}
\usage{
getBondTerm.TFutures(maturity, ticker, decade = "auto", forCFactor = FALSE)
}
\arguments{
\item{maturity}{Maturity date}

\item{ticker}{TFutures ticker}

\item{decade}{Decade that can be "auto" - default value, "pres" - present decade,
"prev" - previous decade, "next" - next decade}
}
\value{
Rounded bond's term in years
}
\description{
Return rounded bond's term to maturity according to TFutures rules
}
