% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{getCarryValue.FIBond}
\alias{getCarryValue.FIBond}
\title{Calculate carry for FIBond object}
\usage{
getCarryValue.FIBond(bond, price, settleDate1, settleDate2, repoRate)
}
\arguments{
\item{bond}{FIBond object (can be a list)}

\item{price}{FIBond clean price in percentage (can be a vector)}

\item{settleDate1}{Start date (can be a vector)}

\item{settleDate2}{End date (can be a vector)}

\item{repoRate}{Funding term repo rate in percentage (can be a vector)}
}
\value{
Carry for FIBond object
}
\description{
Calculate carry for FIBond object
}
