% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getContractType.TFutures}
\alias{getContractType.TFutures}
\title{Return contract type by TFutures ticker}
\usage{
getContractType.TFutures(ticker)
}
\arguments{
\item{ticker}{Ticker (can be a vector)}
}
\value{
Contract type
}
\description{
Return contract type by TFutures ticker
}
