% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIBond.R
\name{getFace.FIBond}
\alias{getFace.FIBond}
\title{Return current face amount}
\usage{
getFace.FIBond(bond, settleDate = nextBizDay())
}
\arguments{
\item{bond}{FIBond object}

\item{settleDate}{Calculation date}
}
\value{
Current face amount
}
\description{
Return current face amount
}
