% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futures.R
\name{getFuturesCodeFromTicker}
\alias{getFuturesCodeFromTicker}
\title{Take contract type code from futures ticker}
\usage{
getFuturesCodeFromTicker(ticker)
}
\arguments{
\item{ticker}{Ticker (can be a vector)}
}
\value{
Contract type code
}
\description{
Take contract type code from futures ticker
}
