% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getIRP.TFututes}
\alias{getIRP.TFututes}
\title{Calculate implied repo rate for TFutures object}
\usage{
getIRP.TFututes(fut, futPrice, bondPrice, tradeDate = Sys.Date(),
  bond = fut$ctd)
}
\arguments{
\item{fut}{TFutures object}

\item{futPrice}{TFutures price in percentage}

\item{bondPrice}{CTD bond clean price in percentage}

\item{tradeDate}{Calculation date}

\item{bond}{FIBond object}
}
\value{
Implied repo rate for TFutures object in percentage
}
\description{
Calculate implied repo rate for TFutures object
}
