% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getIRP.TFututes}
\alias{getIRP.TFututes}
\title{Calculate implied repo rate for TFutures object}
\usage{
getIRP.TFututes(fut, futPrice, bondPrice, tradeDate = Sys.Date(),
  bond = takeCTD.TFutures(fut))
}
\arguments{
\item{fut}{TFutures object (can be a list)}

\item{futPrice}{TFutures price in percentage (can be a vector)}

\item{bondPrice}{CTD bond clean price in percentage (can be a vector)}

\item{tradeDate}{Calculation date (can be a vector)}

\item{bond}{FIBond object (can be a list)}
}
\value{
Implied repo rate for TFutures object in percentage
}
\description{
Calculate implied repo rate for TFutures object
}
