% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futures.R
\name{getMonthNumberFromFuturesTicker}
\alias{getMonthNumberFromFuturesTicker}
\title{Take expiration month number from futures ticker}
\usage{
getMonthNumberFromFuturesTicker(ticker)
}
\arguments{
\item{ticker}{Ticker}
}
\value{
Expiration month number
}
\description{
Take expiration month number from futures ticker
}
