% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{getMonthNumberFromMonthCode}
\alias{getMonthNumberFromMonthCode}
\title{Return month number for a given month futures code}
\usage{
getMonthNumberFromMonthCode(code)
}
\arguments{
\item{code}{Month code}
}
\value{
Month number
}
\description{
Return month number for a given month futures code
}
