% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getNetBasis.TFutures}
\alias{getNetBasis.TFutures}
\title{Return Net Basis for TFutures object and given FIBond}
\usage{
getNetBasis.TFutures(fut, futPrice, bondPrice, repoRate,
  tradeDate = Sys.Date(), bond = fut$ctd)
}
\arguments{
\item{fut}{TFutures object}

\item{futPrice}{TFutures price in percentage}

\item{bondPrice}{FIBond price in percentage}

\item{repoRate}{FIBond repo rate in percentage}

\item{tradeDate}{Trade date}

\item{bond}{FIBond object}
}
\value{
Net Basis in percentage of FIBond face
}
\description{
Return Net Basis for TFutures object and given FIBond
}
