% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getPVBP.FIBond}
\alias{getPVBP.FIBond}
\title{Calculate PVBP of FIBond object}
\usage{
getPVBP.FIBond(bond, price, settleDate = nextBizDay())
}
\arguments{
\item{bond}{FIBond object}

\item{price}{FIBond clean price where 1 is 100\% (can be a vector)}

\item{settleDate}{Calculation date (can be a vector)}
}
\value{
PVBP of FIBond object
}
\description{
Calculate PVBP of FIBond object
}
