% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getPVBP.TFutures}
\alias{getPVBP.TFutures}
\title{Calculate PVBP of TFutures object relative to CTD yield change}
\usage{
getPVBP.TFutures(fut, ctdPrice, repoRate, tradeDate = Sys.Date())
}
\arguments{
\item{fut}{TFutures object}

\item{ctdPrice}{CTD bond clean price for 100 face}

\item{repoRate}{CTD repo rate}

\item{tradeDate}{Calculation date}
}
\value{
PVBP of TFutures object relative to CTD yield change
}
\description{
Calculate PVBP of TFutures object relative to CTD yield change
}
