% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getPVBPRP.TFutures}
\alias{getPVBPRP.TFutures}
\title{Calculate PVBP of TFutures object relative to CTD repo rate change}
\usage{
getPVBPRP.TFutures(fut, ctdPrice, repoRate, tradeDate = Sys.Date())
}
\arguments{
\item{fut}{TFutures object (can be a list)}

\item{ctdPrice}{CTD bond clean price in percentage (can be a vector)}

\item{repoRate}{CTD repo rate in percentage (can be a vector)}

\item{tradeDate}{Calculation date (can be a vector)}
}
\value{
PVBP of TFutures object relative to CTD repo rate change in percentage
}
\description{
Calculate PVBP of TFutures object relative to CTD repo rate change
}
