% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFutures.R
\name{getValue.TFutures}
\alias{getValue.TFutures}
\title{Return value of TFutures contract}
\usage{
getValue.TFutures(fut, futPrice, settlePrice, side = "long")
}
\arguments{
\item{fut}{TFutures object}

\item{futPrice}{TFutures price in percentage}

\item{settlePrice}{TFutures settlement price in percentage}

\item{side}{Side - "long" or "short"}
}
\value{
Value in dollars of one contract
}
\description{
Return value of TFutures contract
}
