% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bond.R
\name{getYield.FIBond}
\alias{getYield.FIBond}
\title{Calculate yield of FIBond object}
\usage{
getYield.FIBond(bond, price, settleDate = nextBizDay(), digits = 5)
}
\arguments{
\item{bond}{FIBond object}

\item{price}{FIBond clean price where 1 is 100\% (can be a vector)}

\item{settleDate}{Calculation date (can be a vector)}

\item{digits}{Accuracy as a number of digits after point}
}
\value{
FIBond yield
}
\description{
Calculate yield of FIBond object
}
