% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{nextBizDay}
\alias{nextBizDay}
\title{Return next business day using RQuantLib calendars}
\usage{
nextBizDay(date = Sys.Date(), calendar = "WeekendsOnly")
}
\arguments{
\item{date}{Date (can be a vector)}

\item{calendar}{RQuantLib calendar name}
}
\value{
Next business day
}
\description{
Return next business day using RQuantLib calendars
}
