% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{nextMonth}
\alias{nextMonth}
\title{Return number of next month}
\usage{
nextMonth(month, year)
}
\arguments{
\item{month}{Number of present month (can be a vector)}

\item{year}{Year (can be a vector)}
}
\value{
List where $month is next month and $year is year of next month
}
\description{
Return number of next month
}
