% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{nextMonth}
\alias{nextMonth}
\title{Return number of next month}
\usage{
nextMonth(month)
}
\arguments{
\item{month}{Number of present month}
}
\value{
Number of next month
}
\description{
Return number of next month
}
