% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{TFutures}
\alias{TFutures}
\title{Constructor of T-bond-note futures contract}
\usage{
TFutures(ticker = NA, ctdFileName = "", dateFormat = "mdy")
}
\arguments{
\item{ticker}{Bloomber ticker like TYU7 etc}

\item{ctdFileName}{name of ctd data file}

\item{dateFormat}{File date format from lubridate package i.e. "dmy", "mdy" etc}
}
\value{
TFutures object
}
\description{
Attributes:
$name - name as you wish
$ticker - ticker like TYU7 etc
$notionalAmount - notional amount of futures
$deliveryDate - Last delivery date that is supposed to be an accual delivary date
$ctd - Ceapest-to-deliver Bond object
}
