% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fval.R
\name{TFutures}
\alias{TFutures}
\title{Constructor of T-bond-note futures contract}
\usage{
TFutures(ticker = NA)
}
\arguments{
\item{ticker}{Ticker or its number from futDataList like TYU7 etc}
}
\value{
TFutures object
}
\description{
Attributes:
$name - name as you wish
$ticker - ticker like TYU7 etc
$notionalAmount - notional amount of futures
$deliveryDate - Delivery date
$ctd - Ceapest-to-deliver Bond object
}
